no convert layernorm

+++ vllm/model_executor/models/llama.py
@@ -191,9 +191,9 @@ class LlamaDecoderLayer(nn.Module):
             linear_method=linear_method,
         )
         self.input_layernorm = RMSNorm(config.hidden_size,
-                                       eps=config.rms_norm_eps).to(torch.float32)
+                                       eps=config.rms_norm_eps)
         self.post_attention_layernorm = RMSNorm(config.hidden_size,
-                                                eps=config.rms_norm_eps).to(torch.float32)
+                                                eps=config.rms_norm_eps)
 
     def forward(
         self,
@@ -206,12 +206,10 @@ class LlamaDecoderLayer(nn.Module):
         # Self Attention
         if residual is None:
             residual = hidden_states
-            hidden_states = self.input_layernorm(hidden_states.to(torch.float32)).to(torch.bfloat16)
+            hidden_states = self.input_layernorm(hidden_states)
         else:
             hidden_states, residual = self.input_layernorm(
-                hidden_states.to(torch.float32), residual.to(torch.float32))
-            hidden_states = hidden_states.to(torch.bfloat16)
-            residual = residual.to(torch.bfloat16)
+                hidden_states, residual)
 
         hidden_states = self.self_attn(
             positions=positions,
@@ -221,9 +219,8 @@ class LlamaDecoderLayer(nn.Module):
         )
         # Fully Connected
         hidden_states, residual = self.post_attention_layernorm(
-            hidden_states.to(torch.float32), residual.to(torch.float32))
-        hidden_states = self.mlp(hidden_states.to(torch.bfloat16))
-        residual =  residual.to(torch.bfloat16)
+            hidden_states, residual)
+        hidden_states = self.mlp(hidden_states)
         return hidden_states, residual
 
 
@@ -251,7 +248,7 @@ class LlamaModel(nn.Module):
             LlamaDecoderLayer(config, linear_method)
             for _ in range(config.num_hidden_layers)
         ])
-        self.norm = RMSNorm(config.hidden_size, eps=config.rms_norm_eps).to(torch.float32)
+        self.norm = RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
 
     def get_input_embeddings(self, input_ids: torch.Tensor) -> torch.Tensor:
         return self.embed_tokens(input_ids)
@@ -278,8 +275,8 @@ class LlamaModel(nn.Module):
                 attn_metadata,
                 residual,
             )
-        hidden_states, _ = self.norm(hidden_states.to(torch.float32), residual.to(torch.float32))
-        return hidden_states.to(torch.bfloat16)
+        hidden_states, _ = self.norm(hidden_states, residual)
+        return hidden_states
 
 
 class LlamaForCausalLM(nn.Module):
